# Improves performance on BindableProperty