--- a/src/Controls/samples/Controls.Sample.UITests/Issues/Issue19329.xaml
+++ b/src/Controls/samples/Controls.Sample.UITests/Issues/Issue19329.xaml
@@ -1,0 +1,16 @@
+﻿<?xml version="1.0" encoding="utf-8" ?>
+<ContentPage xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
+             xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
+             x:Class="Maui.Controls.Sample.Issues.Issue19329">
+    <VerticalStackLayout x:Name="layout" Margin="75,75,75,75" BackgroundColor="LightGray">
+        <Label Text="Tap the bottom red label. The position of the tap should be calculated relative to the upper blue box." />
+        <VerticalStackLayout HeightRequest="80" WidthRequest="80">
+            <BoxView HorizontalOptions="Start" x:Name="UpperBox" AutomationId="UpperBox" Color="CornflowerBlue" WidthRequest="20" HeightRequest="20"/>
+            <Label HorizontalOptions="End" x:Name="TapHere" Text="Tap Here" AutomationId="TapHere" Background="Red" WidthRequest="20" HeightRequest="20">
+                <Label.GestureRecognizers>
+                    <TapGestureRecognizer Tapped="OnTapped" />
+                </Label.GestureRecognizers>
+            </Label>
+        </VerticalStackLayout>
+    </VerticalStackLayout>
+</ContentPage>