--- a/src/Controls/src/Core/Compatibility/Handlers/ListView/Android/SwitchCellRenderer.cs
+++ b/src/Controls/src/Core/Compatibility/Handlers/ListView/Android/SwitchCellRenderer.cs
@@ -19,6 +19,13 @@ namespace Microsoft.Maui.Controls.Handlers.Compatibility
 		{
 			var cell = (SwitchCell)Cell;
 
+			if (item?.Parent is TableView && item.Handler?.PlatformView is SwitchCellView switchCellView)
+			{
+				// TableView doesn't use convertView
+				_view = switchCellView;
+				return _view;
+			}
+
 			if ((_view = convertView as SwitchCellView) == null)
 				_view = new SwitchCellView(context, item);