--- a/src/Controls/src/Core/Platform/GestureManager/GesturePlatformManager.iOS.cs
+++ b/src/Controls/src/Core/Platform/GestureManager/GesturePlatformManager.iOS.cs
@@ -180,7 +180,7 @@ namespace Microsoft.Maui.Controls.Platform
 			var eventTracker = weakEventTracker.Target as GesturePlatformManager;
 			var virtualView = eventTracker?._handler?.VirtualView as View;
 			var platformRecognizer = weakPlatformRecognizer?.Target as UIGestureRecognizer;
-			var platformView = eventTracker?.PlatformView;
+			var platformView = element?.ToPlatform();
 
 			if (virtualView == null)
 				return null;