--- a/src/TestUtils/src/DeviceTests/AssertHelpers.cs
+++ b/src/TestUtils/src/DeviceTests/AssertHelpers.cs
@@ -27,6 +27,27 @@ namespace Microsoft.Maui.DeviceTests
 			}
 		}
 
+		public static async Task AssertEventually(Func<Task<bool>> assertion, int timeout = 1000, int interval = 100, string message = "Assertion timed out")
+		{
+			do
+			{
+				if (await assertion())
+				{
+					return;
+				}
+
+				await Task.Delay(interval);
+				timeout -= interval;
+
+			}
+			while (timeout >= 0);
+
+			if (! (await assertion()))
+			{
+				throw new XunitException(message);
+			}
+		}
+
 		public static async Task<bool> Wait(Func<bool> predicate, int timeout = 1000, int interval = 100) 
 		{
 			do