… .config/dotnet-tools.json
… eng/README.md
… eng/Version.Details.xml
… src/Controls/src/Core/Compatibility/Handlers/Shell/iOS/ShellFlyoutLayoutManager.cs
… src/Controls/src/Core/Compatibility/Handlers/TableView/Android/TableViewModelRenderer.cs
… src/Controls/src/Core/Interactivity/TriggerBase.cs
… src/Controls/src/Core/PublicAPI/net-android/PublicAPI.Unshipped.txt
… src/Controls/src/Core/SetterSpecificity.cs
… src/Controls/src/Core/SetterSpecificityList.cs
… src/Controls/src/Core/VisualStateManager.cs
… src/Controls/tests/Core.UnitTests/SetterSpecificityListTests.cs
… src/Controls/tests/DeviceTests/ControlsHandlerTestBase.cs
… src/Controls/tests/DeviceTests/Elements/Border/BorderTests.cs
… src/Controls/tests/DeviceTests/Elements/CollectionView/CollectionViewTests.Windows.cs
… src/Controls/tests/DeviceTests/Elements/CollectionView/CollectionViewTests.cs
… src/Controls/tests/DeviceTests/Elements/CollectionView/CollectionViewTests.iOS.cs
… src/Controls/tests/DeviceTests/Elements/Entry/EntryTests.iOS.cs
… src/Controls/tests/DeviceTests/Elements/FlyoutPage/FlyoutPageTests.cs
… src/Controls/tests/DeviceTests/Elements/FlyoutPage/FlyoutPageTests.iOS.cs
… src/Controls/tests/DeviceTests/Elements/Frame/FrameTests.iOS.cs
… src/Controls/tests/DeviceTests/Elements/Image/ImageTests.Android.cs
… src/Controls/tests/DeviceTests/Elements/Image/ImageTests.cs
… src/Controls/tests/DeviceTests/Elements/Image/ImageTests.iOS.cs
… src/Controls/tests/DeviceTests/Elements/ListView/ListViewTests.cs
… src/Controls/tests/DeviceTests/Elements/Modal/ModalTests.cs
… src/Controls/tests/DeviceTests/Elements/NavigationPage/NavigationPageTests.cs
… src/Controls/tests/DeviceTests/Elements/ScrollView/ScrollViewTests.cs
… src/Controls/tests/DeviceTests/Elements/Shape/ShapeTests.cs
… src/Controls/tests/DeviceTests/Elements/Shell/ShellFlyoutHeaderBehaviorAndContentTestCases.cs
… src/Controls/tests/DeviceTests/Elements/Shell/ShellFlyoutHeaderScrollTestCases.cs