@albyrock87 2023-10-06
@StephaneDelcroix @jonathanpeppers I've updated PR with some more improvements on the binding part and a general cleanup + documentation on the `SetterSpecificity` struct.