@albyrock87 2023-11-22
@samhouts I've fixed the merge conflicts, so you can remove the `stale` tag

Now that .NET8 GA is out and you have more time, may I ask for a final review on this in order to avoid fixing merge conflicts again? @MartyIX @StephaneDelcroix @jonathanpeppers @mattleibow 

Thanks

Note: grid layout tests are failing, but that's happening also on the `main` branch