@StephaneDelcroix 2023-11-23
I like this idea, it was actually part of the plan, but was easier to implement, test, and fix using a struct instead of a flag.

In the meantime, I've found this API https://learn.microsoft.com/en-us/dotnet/api/system.windows.basevaluesource?view=windowsdesktop-7.0, and if we're reviewing the SetterSpecificity, we should consider that as well.

I'm glad this isn't a public API, and we're allowed to change and experiment