@albyrock87 2023-11-23
@StephaneDelcroix I've updated PR to also handle https://github.com/dotnet/maui/issues/18103 which was fixed by you in the mean time.

Using an enum like `System.Windows.BaseValueSource` to specify the source is for sure an improvement but it requires a lot more refactoring given that all unit tests are using the current constructor.
To ensure no-regression quality of my work I chose to avoid touching existing unit tests.

Sounds good? I mean, can we consider this PR approved or is there other work I need to do?

Thanks