@jonathanpeppers 2023-11-28
@albyrock87 just so you aren't left here waiting for a merge/review. It feels like this change needs to go into .NET 9, so we have a preview cycle to test out this change.

I will bring up this change to review in a future meeting. Expect it to take a little bit, but thanks for the contribution!