@jonathanpeppers 2024-01-10
FYI I changed this PR to target the `net9.0` branch instead.